Link Spotify Playlist functionality to the backend
